+++ src/main/java/dev/wateralt/mc/weapontroll/energy/EnergySources.java
@@ -4,4 +4,6 @@ public class EnergySources {
   public static final int SWEET_BERRIES = 5;
   public static final int INVIS_POT = 50;
   public static final int SPONGE_CAKE = 500;
+  public static final int TRANSFER_PERCENT_PER_S = 10;
+  public static final int STEAL_PERCENT_PER_S = 1;
 }